prompts: inline conversation.py into review.py
31-line single-function file. Move build_conversation_prompt into review.py
(it already imported _get_template_env from there) and drop the standalone module.
