+++ src/mira/github_app/handlers.py
@@ -16,7 +16,7 @@ from mira.core.engine import ReviewEngine
 from mira.dashboard.models_config import llm_config_for
 from mira.github_app.auth import GitHubAppAuth
 from mira.index.store import IndexStore
-from mira.llm.prompts.conversation import build_conversation_prompt
+from mira.llm.prompts.review import build_conversation_prompt
 from mira.llm.provider import SUBMIT_THREAD_REPLY_TOOL, LLMProvider
 from mira.models import PRInfo
 from mira.providers import create_provider
